feat: completed the binary search problem
- It has two pointers one at the start one at the end and each time they
calculate the middel index and it checks if the num at the middle index
is less than the target, the target is on the right and we point the
left pointer to the middle index + 1,
- Else if the number at middle index is more than the target, the target
will be on the left so we point the right pointer to middle index - 1.
- Else if the number at middle index is equal to target it returns the
middle index.
- If the left pointer is bigger than the right one it exits the loop and
it returns -1