feat: completed the remove duplicated from sorted list problem
- It checks if the value of the current node and the one from the next
node are equal if not we move the current pointer to the next node else
it puts the next of the next node as the new next pointer for the
current node.
- After pointing the next pointer to the new node we restart the check
from the current pointer without updating it in case the new next
pointer is equal to the old one we removed.